+++ src/components/index.js
@@ -2,13 +2,11 @@ import '@mdi/font/css/materialdesignicons.min.css';
 import 'vuetify/dist/vuetify.min.css';
 
 import Authentication from './Authentication/';
-import Breadcrumb from './Breadcrumb.vue';
-import FileBrowser from './FileBrowser.vue';
+import DataBrowser from './DataBrowser.vue';
 import Upload from './Upload.vue';
 
 export {
   Authentication,
-  Breadcrumb,
-  FileBrowser,
+  DataBrowser,
   Upload,
 };
